requirements.txt
celery==4.2.1
Django==2.1.3
django-celery==3.2.2
django-tastypie==0.14.2
